+++ packages/react-examples/src/react-charting/VerticalStackedBarChart/VerticalStackedBarChart.CustomAccessibility.Example.tsx
@@ -49,7 +49,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blue,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '61%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 1-1 of 4, 2020/04/30 40%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 1-1 of 4, 2020/04/30 Metadata1 61%' },
       },
       {
         legend: 'Metadata2',
@@ -57,7 +57,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blueMid,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '8%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 1-2 of 4, 2020/04/30 5%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 1-2 of 4, 2020/04/30 Metadata2 8%' },
       },
       {
         legend: 'Metadata3',
@@ -65,7 +65,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blueLight,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '31%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 1-3 of 4, 2020/04/30 20%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 1-3 of 4, 2020/04/30 Metadata3 31%' },
       },
     ];
 
@@ -76,7 +76,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blue,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '33%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 2-1 of 4, 2020/04/30 30%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 2-1 of 4, 2020/04/30 Metadata1 33%' },
       },
       {
         legend: 'Metadata2',
@@ -84,7 +84,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blueMid,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '22%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 2-2 of 4, 2020/04/30 20%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 2-2 of 4, 2020/04/30 Metadata2 22%' },
       },
       {
         legend: 'Metadata3',
@@ -92,7 +92,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blueLight,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '45%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 2-3 of 4, 2020/04/30 40%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 2-3 of 4, 2020/04/30 Metadata3 45%' },
       },
     ];
 
@@ -103,7 +103,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blue,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '43%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 3-1 of 4, 2020/04/30 44%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 3-1 of 4, 2020/04/30 Metadata1 43%' },
       },
       {
         legend: 'Metadata2',
@@ -111,7 +111,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blueMid,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '27%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 3-2 of 4, 2020/04/30 28%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 3-2 of 4, 2020/04/30 Metadata2 27%' },
       },
       {
         legend: 'Metadata3',
@@ -119,7 +119,7 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         color: DefaultPalette.blueLight,
         xAxisCalloutData: '2020/04/30',
         yAxisCalloutData: '30%',
-        callOutAccessibilityData: { ariaLabel: 'Bar series 3-3 of 4, 2020/04/30 30%' },
+        callOutAccessibilityData: { ariaLabel: 'Bar series 3-3 of 4, 2020/04/30 Metadata3 30%' },
       },
     ];
 
@@ -134,7 +134,11 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
             { y: 10, legend: 'Recommended Builds', color: DefaultPalette.redDark },
           ],
         }),
-        stackCallOutAccessibilityData: { ariaLabel: 'Bar stack series 1 of 6' },
+        stackCallOutAccessibilityData: {
+          ariaLabel:
+            'Bar stack series 1 of 3, 0 MetaDate1 61% MetaData2 8% MetaDate3 31% ' +
+            'Recommended Builds 10 Supported Builds 42',
+        },
       },
       {
         chartData: secondChartPoints,
@@ -142,7 +146,9 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
         ...(showLine && {
           lineData: [{ y: 33, legend: 'Supported Builds', color: DefaultPalette.magentaLight }],
         }),
-        stackCallOutAccessibilityData: { ariaLabel: 'Bar stack series 2 of 6' },
+        stackCallOutAccessibilityData: {
+          ariaLabel: 'Bar stack series 2 of 3, 20 MetaDate1 33% MetaData2 22% MetaDate3 45% ' + 'Supported Builds 33',
+        },
       },
       {
         chartData: thirdChartPoints,
@@ -153,7 +159,11 @@ export class VerticalStackedBarChartCustomAccessibilityExample extends React.Com
             { y: 20, legend: 'Recommended Builds', color: DefaultPalette.redDark },
           ],
         }),
-        stackCallOutAccessibilityData: { ariaLabel: 'Bar stack series 3 of 6' },
+        stackCallOutAccessibilityData: {
+          ariaLabel:
+            'Bar stack series 3 of 3, 40 MetaDate1 43% MetaData 27% MetaDate3 30% ' +
+            'Recommended Builds 20 Supported Builds 60',
+        },
       },
     ];