+++ packages/react-components/react-calendar-compat/CHANGELOG.md
@@ -1,8 +1,39 @@
 # Change Log - @fluentui/react-calendar-compat
 
-This log was last generated on Thu, 02 May 2024 11:36:31 GMT and should not be manually modified.
+This log was last generated on Thu, 16 May 2024 09:25:13 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.1.7](https://github.com/microsoft/fluentui/tree/@fluentui/react-calendar-compat_v0.1.7)
+
+Thu, 16 May 2024 09:25:13 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-calendar-compat_v0.1.6..@fluentui/react-calendar-compat_v0.1.7)
+
+### Patches
+
+-  chore: Upgrade react-icons version to 2.0.239 to pick up provider export map fix. ([PR #31287](https://github.com/microsoft/fluentui/pull/31287) by ololubek@microsoft.com)
+- fix: Make header button in overlaid month picker match the other header button's styles. ([PR #31381](https://github.com/microsoft/fluentui/pull/31381) by estebanmu@microsoft.com)
+
+## [0.1.6](https://github.com/microsoft/fluentui/tree/@fluentui/react-calendar-compat_v0.1.6)
+
+Thu, 09 May 2024 19:35:12 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-calendar-compat_v0.1.5..@fluentui/react-calendar-compat_v0.1.6)
+
+### Patches
+
+- Bump @fluentui/react-tabster to v9.21.2 ([PR #31321](https://github.com/microsoft/fluentui/pull/31321) by beachball)
+
+## [0.1.5](https://github.com/microsoft/fluentui/tree/@fluentui/react-calendar-compat_v0.1.5)
+
+Mon, 06 May 2024 12:55:02 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-calendar-compat_v0.1.4..@fluentui/react-calendar-compat_v0.1.5)
+
+### Patches
+
+- Bump @fluentui/react-jsx-runtime to v9.0.37 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-shared-contexts to v9.18.0 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-tabster to v9.21.1 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-utilities to v9.18.8 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
 
 ## [0.1.4](https://github.com/microsoft/fluentui/tree/@fluentui/react-calendar-compat_v0.1.4)