+++ packages/react-components/react-provider/CHANGELOG.md
@@ -1,8 +1,39 @@
 # Change Log - @fluentui/react-provider
 
-This log was last generated on Thu, 02 May 2024 11:36:36 GMT and should not be manually modified.
+This log was last generated on Thu, 16 May 2024 09:25:17 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.15.2](https://github.com/microsoft/fluentui/tree/@fluentui/react-provider_v9.15.2)
+
+Thu, 16 May 2024 09:25:17 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-provider_v9.15.1..@fluentui/react-provider_v9.15.2)
+
+### Patches
+
+-  chore: Upgrade react-icons version to 2.0.239 to pick up provider export map fix. ([PR #31287](https://github.com/microsoft/fluentui/pull/31287) by ololubek@microsoft.com)
+
+## [9.15.1](https://github.com/microsoft/fluentui/tree/@fluentui/react-provider_v9.15.1)
+
+Thu, 09 May 2024 19:35:12 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-provider_v9.15.0..@fluentui/react-provider_v9.15.1)
+
+### Patches
+
+- Bump @fluentui/react-tabster to v9.21.2 ([PR #31321](https://github.com/microsoft/fluentui/pull/31321) by beachball)
+
+## [9.15.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-provider_v9.15.0)
+
+Mon, 06 May 2024 12:55:00 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-provider_v9.14.1..@fluentui/react-provider_v9.15.0)
+
+### Minor changes
+
+- feat: adds TagPicker to custom style hooks ([PR #31272](https://github.com/microsoft/fluentui/pull/31272) by bernardo.sunderhus@gmail.com)
+- Bump @fluentui/react-shared-contexts to v9.18.0 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-tabster to v9.21.1 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-utilities to v9.18.8 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.0.37 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
 
 ## [9.14.1](https://github.com/microsoft/fluentui/tree/@fluentui/react-provider_v9.14.1)