+++ packages/react-components/react-timepicker-compat/CHANGELOG.md
@@ -1,8 +1,49 @@
 # Change Log - @fluentui/react-timepicker-compat
 
-This log was last generated on Thu, 02 May 2024 11:36:44 GMT and should not be manually modified.
+This log was last generated on Thu, 16 May 2024 09:25:20 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.2.14](https://github.com/microsoft/fluentui/tree/@fluentui/react-timepicker-compat_v0.2.14)
+
+Thu, 16 May 2024 09:25:20 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-timepicker-compat_v0.2.13..@fluentui/react-timepicker-compat_v0.2.14)
+
+### Patches
+
+- Bump @fluentui/react-combobox to v9.11.4 ([PR #31390](https://github.com/microsoft/fluentui/pull/31390) by beachball)
+- Bump @fluentui/react-field to v9.1.65 ([PR #31390](https://github.com/microsoft/fluentui/pull/31390) by beachball)
+
+## [0.2.13](https://github.com/microsoft/fluentui/tree/@fluentui/react-timepicker-compat_v0.2.13)
+
+Mon, 13 May 2024 12:34:20 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-timepicker-compat_v0.2.12..@fluentui/react-timepicker-compat_v0.2.13)
+
+### Patches
+
+- Bump @fluentui/react-combobox to v9.11.3 ([PR #31328](https://github.com/microsoft/fluentui/pull/31328) by beachball)
+
+## [0.2.12](https://github.com/microsoft/fluentui/tree/@fluentui/react-timepicker-compat_v0.2.12)
+
+Thu, 09 May 2024 19:35:12 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-timepicker-compat_v0.2.11..@fluentui/react-timepicker-compat_v0.2.12)
+
+### Patches
+
+- Bump @fluentui/react-combobox to v9.11.2 ([PR #31321](https://github.com/microsoft/fluentui/pull/31321) by beachball)
+
+## [0.2.11](https://github.com/microsoft/fluentui/tree/@fluentui/react-timepicker-compat_v0.2.11)
+
+Mon, 06 May 2024 12:55:02 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-timepicker-compat_v0.2.10..@fluentui/react-timepicker-compat_v0.2.11)
+
+### Patches
+
+- Bump @fluentui/react-combobox to v9.11.1 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-field to v9.1.64 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.0.37 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-shared-contexts to v9.18.0 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-utilities to v9.18.8 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
 
 ## [0.2.10](https://github.com/microsoft/fluentui/tree/@fluentui/react-timepicker-compat_v0.2.10)